+++ modes/basic-test-mode/src/index.ts
@@ -83,7 +83,6 @@ function modeFactory() {
         'MeasurementTools',
         'Zoom',
         'WindowLevel',
-        'WindowLevelRegion',
         'Pan',
         'Capture',
         'Layout',