+++ modes/longitudinal/src/index.ts
@@ -90,6 +90,7 @@ function modeFactory({ modeConfiguration }) {
       //   },
       // ]);
 
+
       // Init Default and SR ToolGroups
       initToolGroups(extensionManager, toolGroupService, commandsManager, this.labelConfig);
 
@@ -100,7 +101,6 @@ function modeFactory({ modeConfiguration }) {
         'Pan',
         'TrackballRotate',
         'WindowLevel',
-        'WindowLevelRegion',
         'Capture',
         'Layout',
         'Crosshairs',