+++ platform/app/src/routes/DataSourceWrapper.tsx
@@ -27,7 +27,8 @@ const areLocationsTheSame = (location0, location1) => {
  * @param {object} props
  * @param {function} props.children - Layout Template React Component
  */
-function DataSourceWrapper(props) {
+function DataSourceWrapper(props: withAppTypes) {
+  const { servicesManager } = props;
   const navigate = useNavigate();
   const { children: LayoutTemplate, ...rest } = props;
   const params = useParams();
@@ -188,14 +189,29 @@ function DataSourceWrapper(props) {
       if (isDataInvalid) {
         getData().catch(e => {
           console.error(e);
+
+          const { configurationAPI, friendlyName } = dataSource.getConfig();
           // If there is a data source configuration API, then the Worklist will popup the dialog to attempt to configure it
           // and attempt to resolve this issue.
-          if (dataSource.getConfig().configurationAPI) {
+          if (configurationAPI) {
             return;
           }
 
-          // No data source configuration API, so navigate to the not found server page.
-          navigate('/notfoundserver', '_self');
+          servicesManager.services.uiModalService.show({
+            title: 'Data Source Connection Error',
+            containerDimensions: 'w-1/2',
+            content: () => {
+              return (
+                <div>
+                  <p className="text-red-600">Error: {e.message}</p>
+                  <p>
+                    Please ensure the following data source is configured correctly or is running:
+                  </p>
+                  <div className="mt-2 font-bold">{friendlyName}</div>
+                </div>
+              );
+            },
+          });
         });
       }
     } catch (ex) {